+++ Editor/Scripts/ChatResponses/IndieBuff_BaseResponseHandler.cs
@@ -9,13 +9,13 @@ namespace IndieBuff.Editor
 {
     public abstract class BaseResponseHandler : IResponseHandler
     {
-        protected readonly IMarkdownParser parser;
+        protected IMarkdownParser parser;
         protected BaseResponseHandler(IMarkdownParser parser)
         {
             this.parser = parser;
         }
 
-        public async Task HandleResponse(string userMessage, VisualElement responseContainer, CancellationToken token)
+        public virtual async Task HandleResponse(string userMessage, VisualElement responseContainer, CancellationToken token)
         {
             var messageContainer = responseContainer.Q<VisualElement>("MessageContainer");
             var messageLabel = messageContainer.Q<TextField>();
@@ -44,8 +44,14 @@ namespace IndieBuff.Editor
             }
             catch (Exception e)
             {
-                Debug.LogError(e.Message);
-                HandleError(responseContainer);
+                if (e.Message == "Error: Insufficient credits")
+                {
+                    HandleInsufficientCredits(responseContainer);
+                }
+                else
+                {
+                    HandleError(responseContainer);
+                }
             }
         }
 
@@ -67,13 +73,33 @@ namespace IndieBuff.Editor
             messageLabel.value = "An error has occured. Please try again.";
             IndieBuff_UserInfo.Instance.responseLoadingComplete?.Invoke();
         }
+        public void HandleInsufficientCredits(VisualElement responseContainer)
+        {
+            var messageContainer = responseContainer.Q<VisualElement>("MessageContainer");
+            var messageLabel = messageContainer.Q<TextField>();
+
+            responseContainer.style.visibility = Visibility.Visible;
+            messageLabel.value = "Not enough credits! Please upgrade or top up your account.";
+            var upgradeButton = responseContainer.Q<Button>("ExecuteButton");
+            upgradeButton.style.display = DisplayStyle.Flex;
+            upgradeButton.SetEnabled(true);
+
+            upgradeButton.text = "Add Credits";
+            upgradeButton.clicked += () =>
+            {
+                Application.OpenURL(IndieBuff_EndpointData.GetFrontendBaseUrl() + "/pricing");
+            };
+
+
+            IndieBuff_UserInfo.Instance.responseLoadingComplete?.Invoke();
+        }
 
         public virtual void Cleanup()
         {
 
         }
 
-        private async Task HandleChatDatabase(string userMessage, string aiMessage, string summaryMessage = "")
+        protected virtual async Task HandleChatDatabase(string userMessage, string aiMessage, string summaryMessage = "")
         {
             IndieBuff_UserInfo.Instance.lastUsedMode = IndieBuff_UserInfo.Instance.currentMode;
             IndieBuff_UserInfo.Instance.lastUsedModel = IndieBuff_UserInfo.Instance.selectedModel;
@@ -87,7 +113,7 @@ namespace IndieBuff.Editor
             await IndieBuff_ConvoHandler.Instance.RefreshConvoList();
         }
 
-        protected async Task HandleResponseMetadata(string userMessage, IMarkdownParser parser)
+        protected virtual async Task HandleResponseMetadata(string userMessage, IMarkdownParser parser)
         {
             int splitIndex = parser.GetFullMessage().LastIndexOf('\n');
             string aiMessage;