+++ Editor/Scripts/ChatResponses/IndieBuff_PrototypeResponseHandler.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Indiebuff.Editor;
 using UnityEngine.UIElements;
 
 namespace IndieBuff.Editor