+++ Editor/Scripts/ChatResponses/IndieBuff_ResponseHandlerFactory.cs
@@ -1,5 +1,4 @@
 using System;
-using Indiebuff.Editor;
 using IndieBUff.Editor;
 using UnityEngine.UIElements;