+++ Editor/Scripts/Controllers/IndieBuff_ApiClient.cs
@@ -74,7 +74,7 @@ namespace IndieBuff.Editor
 
         public Task<HttpResponseMessage> LogoutAsync(string refreshToken)
         {
-            IndieBuff_UserInfo.Instance.selectedModel = "Base Model";
+            IndieBuff_UserInfo.Instance.selectedModel = "claude-3-5-sonnet";
             var request = new HttpRequestMessage(HttpMethod.Post, "plugin-auth/logout");
             request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", refreshToken);
             return client.SendAsync(request);
@@ -90,7 +90,7 @@ namespace IndieBuff.Editor
 
             await TokenManager.Instance.RefreshTokensAsync();
             string contextString = await IndieBuff_ContextDriver.Instance.BuildAllContext(prompt);
-            var requestData = new ChatRequest { prompt = prompt, aiModel = IndieBuff_UserInfo.Instance.selectedModel, chatMode = IndieBuff_UserInfo.Instance.currentMode.ToString(), context = contextString, gameEngine = "unity", lastModel = IndieBuff_UserInfo.Instance.lastUsedModel };
+            var requestData = new ChatRequest { prompt = prompt, chatMode = IndieBuff_UserInfo.Instance.currentMode.ToString(), context = contextString, gameEngine = "unity" };
             List<MessageHistoryObject> messageHistory = IndieBuff_ConvoHandler.Instance.currentMessages.Select(message => new MessageHistoryObject
             {
                 role = message.Role,
@@ -126,11 +126,18 @@ namespace IndieBuff.Editor
                         }
                     }
                 }
+                _ = IndieBuff_UserInfo.Instance.GetCredits();
             }
             else
             {
-                Debug.Log(response.Content.ReadAsStringAsync().Result);
-                throw new Exception("Error: Response was unsuccessful");
+                if (response.Content.ReadAsStringAsync().Result.Contains("Insufficient credits"))
+                {
+                    throw new Exception("Error: Insufficient credits");
+                }
+                else
+                {
+                    throw new Exception("Error: Response was unsuccessful");
+                }
             }
         }
 
@@ -139,9 +146,9 @@ namespace IndieBuff.Editor
             return SendRequestAsync(() => client.GetAsync("plugin-chat/user-info"));
         }
 
-        public Task<HttpResponseMessage> GetAvailableModelsAsync()
+        public Task<HttpResponseMessage> GetCreditsAsync()
         {
-            return SendRequestAsync(() => client.GetAsync("plugin-chat/available-models"));
+            return SendRequestAsync(() => client.GetAsync("plugin-chat/credits"));
         }
 
         [Serializable]
@@ -152,8 +159,6 @@ namespace IndieBuff.Editor
             public List<MessageHistoryObject> history = new List<MessageHistoryObject>();
             public string gameEngine;
             public string chatMode;
-            public string aiModel;
-            public string lastModel;
         }
 
         [Serializable]