+++ Editor/Scripts/Models/IndieBuff_Credits.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace IndieBuff.Editor
+{
+    [Serializable]
+    public class IndieBuff_Credits
+    {
+        public int credits;
+        public int topUpCredits;
+    }
+}